requirements.txt
Pillow>=10.0.0
requests>=2.31.0
beautifulsoup4>=4.12.3
